Ok, here's a fix for the -S prog now working correctly with store description
Also, something I discussed before, I bias the rank for OR results. Totally random
ranking -- someday that ranking code will get a review.

Minor tweak for a warning.

+++ src/docprop.c
@@ -557,7 +557,7 @@ static int EncodeProperty( struct metaEntry *meta_entry, char **encodedStr, char
         num = PACKLONG(num);
         tmpnum = (char *)&num;
 
-        for ( j=0; j <= sizeof(num)-1; j++ )
+        for ( j=0; j <= (int)sizeof(num)-1; j++ )
             newstr[j] = (unsigned char)tmpnum[j];
         
         newstr[ sizeof(num) ] = '\0';
